Resolved URL bug when syncing across all blogs
The URL prefix for the Learning Locker activity request was being
cached across all blog_ids.  Now, a new string is generated for each
unique blog_id that runs the sync in the multisite install.

Added a debug flag that can be manually flipped to turn on debug
log statements, provided that WP_DEBUG is TRUE.

Returned to the starting blog_id after syncing.

Updated to version 0.2.4